+++ examples/data-sources/netskope_device_classification_on_prem_detection_list/data-source.tf
@@ -1,0 +1,4 @@
+data "netskope_device_classification_on_prem_detection_list" "my_deviceclassificationonpremdetectionlist" {
+  limit  = 7
+  offset = 1
+}