+++ examples/data-sources/netskope_device_classification_rule_list/data-source.tf
@@ -1,0 +1,5 @@
+data "netskope_device_classification_rule_list" "my_deviceclassificationrulelist" {
+  label  = "...my_label..."
+  limit  = 5
+  offset = 5
+}