+++ examples/resources/netskope_device_classification_on_prem_detection/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = netskope_device_classification_on_prem_detection.my_netskope_device_classification_on_prem_detection
+  id = 0
+}