+++ examples/resources/netskope_device_classification_rule/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = netskope_device_classification_rule.my_netskope_device_classification_rule
+  id = 0
+}