+++ examples/resources/netskope_device_classification_steering_mapping/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = netskope_device_classification_steering_mapping.my_netskope_device_classification_steering_mapping
+  id = 0
+}