+++ examples/resources/netskope_device_classification_steering_mapping/import.sh
@@ -1,0 +1,1 @@
+terraform import netskope_device_classification_steering_mapping.my_netskope_device_classification_steering_mapping 0