+++ internal/provider/testdata/TestAccDeviceClassificationRuleListDataSource_basic/main.tf
@@ -1,0 +1,35 @@
+variable "name" {
+  type = string
+}
+
+resource "netskope_device_classification_tag" "test" {
+  name = "${var.name}-tag"
+}
+
+resource "netskope_device_classification_rule" "test" {
+  name  = var.name
+  label = netskope_device_classification_tag.test.name
+  os    = "windows"
+
+  conditions = jsonencode({
+    "$and" = [
+      {
+        "$and" = [
+          {
+            "$and" = [
+              {
+                "min_os_version_check" = {
+                  "min_os_version" = "10.0.0"
+                }
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  })
+}
+
+data "netskope_device_classification_rule_list" "test" {
+  depends_on = [netskope_device_classification_rule.test]
+}